Move PrecinctScannerState to libs/types
