+++ libs/types/src/index.ts
@@ -17,6 +17,7 @@ export * from './image';
 export * from './interpretation';
 export * from './numeric';
 export * from './polls';
+export * from './precinct_scanner';
 export * from './precinct_selection';
 export * from './printing';
 export * from './system_settings';
